+++ component_catalog/views.py
@@ -29,6 +29,7 @@ from django.http import HttpResponse
 from django.http import JsonResponse
 from django.shortcuts import get_object_or_404
 from django.shortcuts import redirect
+from django.shortcuts import render
 from django.urls import reverse
 from django.urls import reverse_lazy
 from django.utils.dateparse import parse_datetime
@@ -39,6 +40,7 @@ from django.utils.translation import gettext
 from django.utils.translation import gettext_lazy as _
 from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.csrf import csrf_protect
+from django.views.decorators.http import require_GET
 from django.views.decorators.http import require_POST
 from django.views.generic import FormView
 from django.views.generic.edit import BaseFormView
@@ -85,6 +87,7 @@ from dje.utils import get_cpe_vuln_link
 from dje.utils import get_help_text as ght
 from dje.utils import get_preserved_filters
 from dje.utils import is_available
+from dje.utils import is_hx_request
 from dje.utils import is_uuid4
 from dje.utils import localized_datetime
 from dje.utils import remove_empty_values
@@ -1456,6 +1459,7 @@ def package_scan_view(request, dataspace, uuid):
     dataspace = user.dataspace
     package = get_object_or_404(Package, uuid=uuid, dataspace=dataspace)
     download_url = package.download_url
+    is_hxr = is_hx_request(request)
 
     scancodeio = ScanCodeIO(dataspace)
     if scancodeio.is_configured() and dataspace.enable_package_scanning:
@@ -1466,13 +1470,71 @@ def package_scan_view(request, dataspace, uuid):
                 user_uuid=user.uuid,
                 dataspace_uuid=dataspace.uuid,
             )
+
+            if is_hxr:
+                template = "product_portfolio/tables/scan_progress_cell.html"
+                scan = scancodeio.get_scan_results(
+                    download_url=download_url,
+                    dataspace=dataspace,
+                )
+                scan["download_result_url"] = get_scan_results_as_file_url(scan)
+
+                status = scancodeio.get_status_from_scan_results(scan)
+                needs_refresh = False
+                if status in ["running", "not_started", "queued"]:
+                    needs_refresh = True
+
+                context = {
+                    "package": package,
+                    "scan": scan,
+                    "view_url": package.get_absolute_url(),
+                    "needs_refresh": needs_refresh,
+                }
+                return render(request, template, context)
+
             scancode_msg = "The Package URL was submitted to ScanCode.io for scanning."
             messages.success(request, scancode_msg)
         else:
+            if is_hxr:
+                return HttpResponse("URL is not reachable.")
+
             scancode_msg = f"The URL {download_url} is not reachable."
             messages.error(request, scancode_msg)
 
+    if is_hxr:
+        return Http404
+
     return redirect(f"{package.details_url}#scan")
+
+
+@login_required
+@require_GET
+def get_scan_progress_htmx_view(request, dataspace, uuid):
+    template = "product_portfolio/tables/scan_progress_cell.html"
+    dataspace = request.user.dataspace
+    package = get_object_or_404(Package, uuid=uuid, dataspace=dataspace)
+    scancodeio = ScanCodeIO(dataspace)
+
+    scan = scancodeio.get_scan_results(
+        download_url=package.download_url,
+        dataspace=dataspace,
+    )
+    if not scan:
+        raise Http404("Scan not found.")
+
+    status = scancodeio.get_status_from_scan_results(scan)
+    needs_refresh = False
+    if status in ["running", "not_started", "queued"]:
+        needs_refresh = True
+
+    scan["download_result_url"] = get_scan_results_as_file_url(scan)
+    context = {
+        "package": package,
+        "scan": scan,
+        "view_url": package.get_absolute_url(),
+        "needs_refresh": needs_refresh,
+    }
+    return render(request, template, context)
 
 
 @login_required